// src/lib/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;

export interface Account {
  accessToken: string | null;
  accessTokenExpiresAt: number | null;
  accountId: string;
  createdAt: Generated<number>;
  id: string;
  idToken: string | null;
  password: string | null;
  providerId: string;
  refreshToken: string | null;
  refreshTokenExpiresAt: number | null;
  scope: string | null;
  updatedAt: number;
  userId: string;
}

export interface Item {
  checked: Generated<number | null>;
  createdAt: Generated<number>;
  id: string;
  listId: string;
  name: string;
  updatedAt: Generated<number>;
}

export interface List {
  createdAt: Generated<number>;
  id: string;
  name: string;
  updatedAt: Generated<number>;
  userId: string | null;
}

export interface Session {
  createdAt: Generated<number>;
  expiresAt: number;
  id: string;
  ipAddress: string | null;
  token: string;
  updatedAt: number;
  userAgent: string | null;
  userId: string;
}

export interface User {
  createdAt: Generated<number>;
  email: string;
  emailVerified: Generated<number>;
  id: string;
  image: string | null;
  name: string;
  updatedAt: Generated<number>;
}

export interface Verification {
  createdAt: Generated<number>;
  expiresAt: number;
  id: string;
  identifier: string;
  updatedAt: Generated<number>;
  value: string;
}

export interface DB {
  account: Account;
  item: Item;
  list: List;
  session: Session;
  user: User;
  verification: Verification;
}